+++ src/firecrest/compute/models.py
@@ -4,6 +4,7 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
 from typing import List, Optional
+from pydantic import Field
 
 # models
 from lib.models import CamelModel
@@ -27,7 +28,11 @@ class PostJobSubmitRequest(JobSubmitRequestModel):
                         "standard_input": "/dev/null",
                         "standard_output": "count_to_100.out",
                         "standard_error": "count_to_100.err",
-                        "script": "#!/bin/bash\nfor i in {1..100}\ndo\necho $i\nsleep 1\ndone",
+                        "env": {
+                            "LD_LIBRARY_PATH": "/path/to/library",
+                            "PATH": "/path/to/bin"
+                        },
+                        "script": "#!/bin/bash\n--partition=part01\nfor i in {1..100}\ndo\necho $i\nsleep 1\ndone",
                     }
                 }
             ]
@@ -48,4 +53,13 @@ class PostJobSubmissionResponse(CamelModel):
 
 
 class PostJobAttachRequest(CamelModel):
-    command: str = None
+    command: str = Field(default=None, description="Command to attach to the job")
+    model_config = {
+        "json_schema_extra": {
+            "examples": [
+                {
+                    "command": "echo 'Attached with success' > $HOME/attach.out"
+                }
+            ]
+        }
+    }