+++ CHANGELOG.md
@@ -10,11 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - JupyterHub example
+- Documentation for logging architecture
+- POST and PUT bodies request examples
 
 ### Changed
 
 - Documentation for logging architecture
 - Images for documentation
+- Description of API definition
 
 ### Fixed